plot_mask_vals in the root plotting.py instead of in tms experiment
